generate-issue.sh: Do not flood console with /etc/issue
The networkmanager hook `generate-issue.sh` is called when
the network interfaces changed, this is how /etc/issue gets
the current IP address. Somehow it is called around 4 times
during boot, which makes /etc/issue flood the console.

This change makes generate-issue.sh able to detect if /etc/issue
changed, and if no, do not send agetty HUP signal to re-display
/etc/issue.

+++ meta-titania/recipes-connectivity/networkmanager/files/generate-issue.sh
@@ -1,17 +1,29 @@
 #!/bin/bash
 
+OLD_ISSUE_HASH=
+if [[ -f /etc/issue ]]
+then
+    OLD_ISSUE_HASH="$(sha256sum /etc/issue)"
+fi
+
 cat /etc/titania.ascii > /etc/issue
 
 IPS=$(ip -4 addr show | grep -w inet | grep -vw lo | grep -wv docker0 | awk '{print $2}' | cut -d/ -f1)
 if [[ $? = 0 ]]
 then
-	for IP in $IPS
-	do
-	    echo -e "\t\t\tTitania Web Interface:  http://$IP/" >> /etc/issue
-	done
+    for IP in $IPS
+    do
+        echo -e "\t\t\tTitania Web Interface:  http://$IP/" >> /etc/issue
+    done
 fi
 
 echo >> /etc/issue
 cat /etc/issue.titania >> /etc/issue
 
-killall -HUP agetty
+NEW_ISSUE_HASH="$(sha256sum /etc/issue)"
+if [[ $NEW_ISSUE_HASH != $OLD_ISSUE_HASH ]]
+then
+    # agetty will show the changed /etc/issue if killed with HUP signal.
+    # If /etc/issue didn't change, no need to kill agetty.
+    killall -HUP agetty
+fi
